Update api "lifts/[type]" and "follows"
- change their api route (no use of UserID)
- add server side session function to fetch current UserID
