+++ .env
@@ -1,2 +1,3 @@
 GITHUB_ID=5eb4074f800b9f69aa44
 GITHUB_SECRET=c249cdd0f2a948b8f74a77d58f72dead261439a8
+AUTH_SECRET=say_fuck_naTi0na1_chia0_tUng_uNiversity
