+++ src/pages/api/lifts/[type].js
@@ -1,10 +1,46 @@
 import mysql from 'mysql2/promise';
+import { authOptions } from '/src/pages/api/auth/[...nextauth]'
+import { getServerSession } from "next-auth/next"
+
+// for session
+// consider store this function in another file (e.g., common.js)?
+export async function getServerSideProps(context) {
+  console.log("getServerSideProps");
+  const session = await getServerSession(context.req, context.res, authOptions)
+
+  if (!session) {
+    return {
+      redirect: {
+        destination: '/',
+        permanent: false,
+      },
+    }
+  }
+
+  return {
+    props: {
+      session,
+    },
+  }
+}
 
 export default async function handler(req, res) {
+  const session = await getServerSession(req, res, authOptions);
+
+  // maybe 改成不要 return false，而是 return data without "isFollow" make more sense?
+  if (!session) {
+    res.status(404).json({
+      success: false,
+      message: 'You must be logged in',
+    });
+    return;
+  }
+
+  console.log(session);
+
   const query = req.query;
   const type = query['type'];
-  const id = parseInt(query['id']);
-  // const page = parseInt(query['page']);
+  const id = parseInt(session.user.id);
 
   if (type !== 'squat' && type !== 'bench' && type !== 'deadlift') {
     // error
